[Chore] Fix merge conflicts
