--- app/src/main/java/com/example/antonyng/level_up/RecyclerViewAdapter.java
@@ -1,85 +1,0 @@
-package com.example.antonyng.level_up;
-
-import android.content.Context;
-import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
-import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.RelativeLayout;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import com.bumptech.glide.Glide;
-
-import java.util.ArrayList;
-
-import de.hdodenhof.circleimageview.CircleImageView;
-
-public class RecyclerViewAdapter extends RecyclerView.Adapter<RecyclerViewAdapter.ViewHolder> {
-    private static final String TAG = "RecyclerViewAdapter";
-
-    private ArrayList<String> mImageNames;
-    private ArrayList<String> mImage;
-    private Context mcontext;
-
-    public RecyclerViewAdapter(Context mcontext, ArrayList<String> mImageNames, ArrayList<String> mImage) {
-        this.mImageNames = mImageNames;
-        this.mImage = mImage;
-        this.mcontext = mcontext;
-    }
-
-    @NonNull
-    @Override
-    public ViewHolder onCreateViewHolder(@NonNull ViewGroup viewGroup, int i) {
-        View view = LayoutInflater.from(viewGroup.getContext()).inflate(R.layout.layout_listitem, viewGroup, false);
-        return new ViewHolder(view);
-    }
-
-    @Override
-    public void onBindViewHolder(@NonNull ViewHolder viewHolder, final int i) {
-        Log.d(TAG, "onBindViewHolder: called");
-
-        Glide.with(mcontext)
-                .asBitmap()
-                .load(mImage.get(i))
-                .into(viewHolder.image);
-
-        viewHolder.imageName.setText(mImageNames.get(i));
-
-        viewHolder.parentLayout.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Log.d(TAG, "onClick: clicked on:" + mImageNames.get(i));
-
-                Toast.makeText(mcontext, mImageNames.get(i), Toast.LENGTH_SHORT).show();
-
-                Intent intent = new Intent(mcontext, DetailActivity.class);
-                intent.putExtra("image_url", mImage.get(i));
-                intent.putExtra("dev_name", mImageNames.get(i));
-                mcontext.startActivity(intent);
-            }
-        });
-    }
-
-    @Override
-    public int getItemCount() {
-        return mImageNames.size();
-    }
-
-    public class ViewHolder extends RecyclerView.ViewHolder{
-
-        CircleImageView image;
-        TextView imageName;
-        RelativeLayout parentLayout;
-
-        public ViewHolder(@NonNull View itemView) {
-            super(itemView);
-            image = itemView.findViewById(R.id.image);
-            imageName = itemView.findViewById(R.id.image_name);
-            parentLayout = itemView.findViewById(R.id.parent_layout);
-        }
-    }
-}
